Mol: split files and format
