+++ psi4/driver/qcdb/molparse/__init__.py
@@ -1,3 +1,4 @@
-from .tofrom import from_arrays, from_string, from_input_arrays
+from .from_arrays import from_arrays, from_input_arrays
+from .from_string import from_string
 from .nucleus import reconcile_nucleus, parse_nucleus_label
 from .chgmult import validate_and_fill_chgmult
