+++ psi4/driver/qcdb/util/__init__.py
@@ -1,4 +1,4 @@
 from .np_blockwise import blockwise_expand, blockwise_contract
 from .np_rand3drot import random_rotation_matrix
 from .mpl import plot_coord
-from .misc import distance_matrix, update_with_error, standardize_efp_angles_units
+from .misc import distance_matrix, update_with_error, standardize_efp_angles_units, filter_comments
